+++ packages/transmute-cli/src/commands/ls.js
@@ -1,0 +1,7 @@
+const run = require('./runner');
+
+export function ls() {
+  var ls_cmd = 'az aks list';
+  console.log(ls_cmd);
+  run.ner( ls_cmd );
+}
